requirements.txt
certifi==2023.11.17
elastic-transport==8.11.0
elasticsearch==8.11.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
urllib3==2.1.0
